+++ src/commands/autocomplete/create.ts
@@ -4,6 +4,7 @@ import * as util from 'util'
 import * as fs from 'fs-extra'
 
 import bashAutocomplete from '../../autocomplete/bash'
+import ZshCompWithSpaces from '../../autocomplete/zsh'
 import bashAutocompleteWithSpaces from '../../autocomplete/bash-spaces'
 import {AutocompleteBase} from '../../base'
 import {Interfaces } from '@oclif/core'
@@ -55,7 +56,14 @@ export default class Create extends AutocompleteBase {
     await fs.writeFile(this.bashSetupScriptPath, this.bashSetupScript)
     await fs.writeFile(this.bashCompletionFunctionPath, this.bashCompletionFunction)
     await fs.writeFile(this.zshSetupScriptPath, this.zshSetupScript)
-    await fs.writeFile(this.zshCompletionFunctionPath, this.zshCompletionWithSpacesFunction)
+
+    // zsh
+    if (this.config.topicSeparator === ":") {
+      await fs.writeFile(this.zshCompletionFunctionPath, this.zshCompletionFunction)
+    } else {
+      const zshCompWithSpaces = new ZshCompWithSpaces(this.config)
+      await fs.writeFile(this.zshCompletionFunctionPath, zshCompWithSpaces.generate())
+    }
   }
 
   private get bashSetupScriptPath(): string {
@@ -202,315 +210,6 @@ compinit;\n`
     return bashScript.replace(/<CLI_BIN>/g, cliBin).replace(/<BASH_COMMANDS_WITH_FLAGS_LIST>/g, this.bashCommandsWithFlagsList)
   }
 
-
-  private get zshCompletionWithSpacesFunction(): string {
-    const argTemplate = `        "%s")\n          %s\n        ;;\n`
-
-    let topics = this.config.topics.filter((topic: Interfaces.Topic) => {
-      // it is assumed a topic has a child if it has children
-      const hasChild = this.config.topics.some(subTopic => subTopic.name.includes(`${topic.name}:`))
-      return hasChild
-    })
-      .sort((a, b) => {
-        if (a.name < b.name) {
-          return -1;
-        }
-        if (a.name > b.name) {
-          return 1;
-        }
-        return 0;
-      })
-      .map(t=> {
-        return {
-          name: t.name,
-          description: sanitizeDescription(t.description)
-        }
-      })
-
-    const coTopics:string[]=[]
-    
-    for (const topic of topics) {
-      for (const cmd of this.commands) {
-        if (topic.name === cmd.id) {
-          coTopics.push(topic.name)
-        }
-      }
-    }
-
-
-    const genZshFlagArgumentsBlock = (flags?: { [name: string]: Interfaces.Command.Flag; }): string => {
-      // if a command doesn't have flags make it only complete files
-      if (!flags) return '_arguments "*: :_files"'
-
-      const flagNames = Object.keys(flags)
-
-      // `-S`:
-      // Do not complete flags after a ‘--’ appearing on the line, and ignore the ‘--’. For example, with -S, in the line:
-      // foobar -x -- -y
-      // the ‘-x’ is considered a flag, the ‘-y’ is considered an argument, and the ‘--’ is considered to be neither.
-      let argumentsBlock = '_arguments -S \\\n'
-
-      for (const flagName of flagNames){
-        const f = flags[flagName]
-
-        // skip hidden flags
-        if (f.hidden) continue
-
-        f.summary = sanitizeDescription(f.summary || f.description)
-
-        let flagSpec = ''
-
-        if (f.type ==='option') {
-          if (f.char) {
-            if (f.multiple) {
-              // this flag can be present multiple times on the line
-              flagSpec += `"*"{-${f.char},--${f.name}}`
-            } else {
-              flagSpec += `"(-${f.char} --${f.name})"{-${f.char},--${f.name}}`
-            }
-
-            flagSpec += `"[${f.summary}]`
-
-            if (f.options) {
-              flagSpec += `:${f.name} options:(${f.options?.join(' ')})"`
-            } else {
-              flagSpec += ':file:_files"'
-            }
-          } else {
-            if (f.multiple) {
-              // this flag can be present multiple times on the line
-              flagSpec += '"*"'
-            }
-
-            flagSpec += `--${f.name}"[${f.summary}]:`
-
-            if (f.options) {
-              flagSpec += `${f.name} options:(${f.options.join(' ')})"`
-            } else {
-              flagSpec += 'file:_files"'
-            }
-          }
-        } else {
-          // Flag.Boolean
-          if (f.char) {
-            flagSpec += `"(-${f.char} --${f.name})"{-${f.char},--${f.name}}"[${f.summary}]"`
-          } else {
-            flagSpec+=`--${f.name}"[${f.summary}]"`
-          }
-        }
-
-        flagSpec += ' \\\n'
-        argumentsBlock += flagSpec
-      }
-      // complete files if `-` is not present on the current line
-      argumentsBlock+='"*: :_files"'
-
-      return argumentsBlock 
-    }
-
-    const genZshValuesBlock = (subArgs: {id: string, summary?: string}[]): string => {
-      let valuesBlock = '_values "completions" \\\n'
-
-      subArgs.forEach(subArg => {
-        valuesBlock += `"${subArg.id}[${subArg.summary}]" \\\n`
-      })
-
-      return valuesBlock
-    }
-
-    const genZshTopicCompFun = (id: string): string => {
-      const flagArgsTemplate = `        "%s")\n          %s\n        ;;\n`
-
-      const underscoreSepId = id.replace(/:/g,'_')
-      const depth = id.split(':').length
-
-      const isCotopic = coTopics.includes(id)
-
-      if (isCotopic) {
-        const compFuncName = `${this.cliBin}_${underscoreSepId}`
-
-        const coTopicCompFunc =
-`_${compFuncName}() {
-  _${compFuncName}_flags() {
-    local context state state_descr line
-    typeset -A opt_args
-  
-    ${genZshFlagArgumentsBlock(this.commands.find(c=>c.id===id)?.flags)}
-  }
-
-  local context state state_descr line
-  typeset -A opt_args
-
-  _arguments -C "1: :->cmds" "*::arg:->args"
-
-  case "$state" in
-    cmds)
-      if [[ "\${words[CURRENT]}" == -* ]]; then
-        _${compFuncName}_flags
-      else
-%s
-      fi
-      ;;
-    args)
-      case $line[1] in
-%s
-      *)
-        _${compFuncName}_flags
-      ;;
-      esac
-      ;;
-  esac 
-}
-`
-        const subArgs: {id: string, summary?: string}[] = []
-
-        let argsBlock = ''
-
-        topics
-          .filter(t => t.name.startsWith(id + ':') && t.name.split(':').length === depth + 1)
-          .forEach(t => {
-            const subArg = t.name.split(':')[depth]
-
-            subArgs.push({
-              id: subArg,
-              summary: t.description
-            })
-
-          argsBlock+= util.format(argTemplate,subArg,`_${this.cliBin}_${underscoreSepId}_${subArg}`) 
-        })
-
-        this.commands
-          .filter(c => c.id.startsWith(id + ':') && c.id.split(':').length === depth + 1)
-          .forEach(c => {
-            if (coTopics.includes(c.id)) return
-            const subArg = c.id.split(':')[depth]
-
-              subArgs.push({
-                id: subArg,
-                summary: c.description
-              })
-
-              argsBlock+= util.format(flagArgsTemplate,subArg,genZshFlagArgumentsBlock(c.flags)) 
-            })
-
-        return util.format(coTopicCompFunc, genZshValuesBlock(subArgs), argsBlock)
-      } else {
-        let argsBlock = ''
-        
-        const subArgs: {id: string, summary?: string}[] = []
-        topics
-          .filter(t => t.name.startsWith(id + ':') && t.name.split(':').length === depth + 1)
-          .forEach(t => {
-            const subArg = t.name.split(':')[depth]
-
-            subArgs.push({
-              id: subArg,
-              summary: t.description
-            })
-
-            argsBlock+= util.format(argTemplate,subArg,`_${this.cliBin}_${underscoreSepId}_${subArg}`) 
-          })
-
-        this.commands
-          .filter(c => c.id.startsWith(id + ':') && c.id.split(':').length === depth + 1)
-          .forEach(c => {
-            if (coTopics.includes(c.id)) return
-            const subArg = c.id.split(':')[depth]
-
-            subArgs.push({
-              id: subArg,
-              summary: c.description
-            })
-
-            argsBlock+= util.format(flagArgsTemplate,subArg,genZshFlagArgumentsBlock(c.flags)) 
-          })
-
-        const topicCompFunc =
-`_${this.cliBin}_${underscoreSepId}() {
-  local context state state_descr line
-  typeset -A opt_args
-
-  _arguments -C "1: :->cmds" "*::arg:->args"
-
-  case "$state" in
-    cmds)
-%s
-      ;;
-    args)
-      case $line[1] in
-%s
-      esac
-      ;;
-  esac 
-}
-`
-        return util.format(topicCompFunc, genZshValuesBlock(subArgs), argsBlock)
-      }
-    }
-
-    const firstArgs: {id: string, summary?: string}[] = []
-
-    topics.forEach(t=>{
-      if(!t.name.includes(':')) firstArgs.push({
-        id: t.name,
-        summary: t.description
-      })
-    })
-    this.commands.forEach(c => { 
-      if(!firstArgs.find(a=> a.id === c.id) && !c.id.includes(':')) firstArgs.push({
-        id: c.id,
-        summary: c.description
-      })
-    })
-
-    const mainArgsCaseBlock = () => {
-      let caseBlock = 'case $line[1] in\n'
-
-      for (const arg of firstArgs) {
-        if (coTopics.includes(arg.id)) {
-          // coTopics already have a completion function.
-          caseBlock +=`${arg.id})\n  _${this.cliBin}_${arg.id} \n  ;;\n`
-        } else {
-          const cmd = this.commands.find(c=>c.id===arg.id)
-
-          // if it's a command and has flags, inline flag completion statement.
-          if (cmd && Object.keys(cmd.flags).length > 0) {
-            caseBlock += `${arg.id})\n${genZshFlagArgumentsBlock(cmd.flags)} ;; \n`
-          } else {
-            // it's a topic, redirect to its completion function.
-            caseBlock +=`${arg.id})\n  _${this.cliBin}_${arg.id} \n  ;;\n`
-          }
-        }
-      }
-
-      caseBlock+='esac\n;;'
-
-      return caseBlock
-    }
-
-    const compFunc =
-`#compdef ${this.cliBin}
-
-${topics.map(t=> genZshTopicCompFun(t.name)).join('\n')}
-
-_${this.cliBin}() {
-  local line state
-
-  _arguments -C "1: :->cmds" "*::arg:->args"
-
-  case "$state" in
-      cmds)
-        ${genZshValuesBlock(firstArgs)} 
-          ;;
-      args)
-        ${mainArgsCaseBlock()}
-  esac
-}
-
-_${this.cliBin}
-`
-  return compFunc
-  }
   private get zshCompletionFunction(): string {
     const cliBin = this.cliBin
     const allCommandsMeta = this.genAllCommandsMetaString
